# apps/api/openapi.yml
openapi: 3.1.0
info:
  title: ParkingMate API
  description: Backend API for ParkingMate parking management system
  version: 1.0.0
  contact:
    name: ParkingMate Team
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.parkingmate.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true

    # Auth Schemas
...
    Customer:
      type: object
      required:
        - id
        - orgId
        - email
        - name
        - role
        - status
        - cars
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        registered:
          type: boolean
        status:
          type: string
        membershipStatus:
          type: string
          enum: [active, expired]
        membershipExpiryDate:
          type: string
          format: date-time
        address:
          type: string
        postalCode:
          type: string
        city:
          type: string
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Car:
      type: object
      required:
        - id
        - orgId
        - licensePlate
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        licensePlate:
          type: string
        brand:
          type: string
        model:
          type: string
        label:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Parking Lot Schemas
...
paths:
...
  /customer/info:
    get:
      tags:
        - users
      summary: Get customer information
      description: Retrieve the authenticated customer profile information
      operationId: getCustomer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer information retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Customer"
    put:
      tags:
        - users
      summary: Update customer information
      description: Update the authenticated customer profile information
      operationId: updateCustomer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Customer information updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/customers:
    get:
      tags:
        - users
      summary: List all customers
      description: Get all customers for the admin's organization
      operationId: getAllCustomers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Customer"
    post:
      tags:
        - users
      summary: Create customer
      description: Create a new customer with optional car data
      operationId: createCustomerByAdmin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
                address:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                isPermanentParker:
                  type: boolean
                  default: false
                parkingStartDate:
                  type: string
                  format: date
                parkingEndDate:
                  type: string
                  format: date
                cars:
                  type: array
                  items:
                    type: object
                    required:
                      - licensePlate
                    properties:
                      licensePlate:
                        type: string
                      brand:
                        type: string
                      model:
                        type: string
                      label:
                        type: string
      responses:
        "201":
          description: Customer created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Customer"
  /admin/customers/{id}:
    get:
      tags:
        - users
      summary: Get customer by ID
      description: Get customer details with all associated cars
      operationId: getCustomerById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Customer"
    put:
      tags:
        - users
      summary: Update customer
      description: Update customer information and cars
      operationId: updateCustomerByAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
                address:
                  type: string
                postalCode:
                  type: string
                city:
                  type: string
                isPermanentParker:
                  type: boolean
                parkingStartDate:
                  type: string
                  format: date
                parkingEndDate:
                  type: string
                  format: date
                cars:
                  type: array
                  items:
                    type: object
                    required:
                      - licensePlate
                    properties:
                      id:
                        type: string
                        format: uuid
                      licensePlate:
                        type: string
                      brand:
                        type: string
                      model:
                        type: string
                      label:
                        type: string
      responses:
        "200":
          description: Customer updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Customer"
    delete:
      tags:
        - users
      summary: Delete customer
      description: Soft delete a customer (sets status to inactive)
      operationId: deleteCustomerByAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Customer deleted successfully

  # ============================================================================
  # Parking Lot Management Endpoints
  # ============================================================================
...